refactor: fix import, exception fix
기존과 바뀐 코드에 맞도록 import 수정, service 로직 수정, delete시 적용방법 수정, 예외처리 관련 위치 수정
