+++ user-service/src/main/java/com/oneultanda/userservice/application/dto/comand/RegisterUserCommand.java
@@ -1,7 +1,5 @@
 package com.oneultanda.userservice.application.dto.comand;
 
-import com.oneultanda.userservice.domain.entity.User;
-
 public record RegisterUserCommand(
         String username,
         String password,
