+++ user-service/src/main/java/com/oneultanda/userservice/application/dto/comand/UpdateUserRoleCommand.java
@@ -1,6 +1,6 @@
 package com.oneultanda.userservice.application.dto.comand;
 
-import com.oneultanda.userservice.domain.entity.Role;
+import com.oneultanda.userservice.domain.model.Role;
 
 public record UpdateUserRoleCommand(
         Role role
