+++ user-service/src/main/java/com/oneultanda/userservice/application/exception/ApplicationErrorCode.java
@@ -1,11 +1,11 @@
-package com.oneultanda.userservice.presentation.exception;
+package com.oneultanda.userservice.application.exception;
 
 import com.oneultanda.userservice.common.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 
 @RequiredArgsConstructor
-public enum PresentationErrorCode implements ErrorCode {
+public enum ApplicationErrorCode implements ErrorCode {
     RESOURCE_NOT_FOUND(HttpStatus.NOT_FOUND, "요청한 리소스를 찾을 수 없습니다"),
     INVALID_REQUEST(HttpStatus.BAD_REQUEST, "잘못된 요청입니다"),
     ACCESS_DENIED(HttpStatus.FORBIDDEN, "접근 권한이 없습니다");
