+++ user-service/src/main/java/com/oneultanda/userservice/presentation/dto/request/UpdateUserRoleRequest.java
@@ -1,7 +1,7 @@
 package com.oneultanda.userservice.presentation.dto.request;
 
 import com.oneultanda.userservice.application.dto.comand.UpdateUserRoleCommand;
-import com.oneultanda.userservice.domain.entity.Role;
+import com.oneultanda.userservice.domain.model.Role;
 
 public record UpdateUserRoleRequest(
         Role role
